A simple vehicle with Bluetooth module that can be controlled via an application
The vehicle can be controlled via an application on an Android phone, which establishes a connection via Bluetooth. It consists of readily available and affordable electronic components in the market. Detailed information about the connections and the purpose of each component is available in the project report I shared. You can learn how the connections were made and what each part does from there.